+++ app/Http/Requests/Testimonial/UpdateTestimonialRequest.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace App\Http\Requests\Testimonial;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class UpdateTestimonialRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     */
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'client_name' => 'required|string|max:255',
+            'company' => 'nullable|string|max:255',
+            'position' => 'nullable|string|max:255',
+            'content' => 'required|string',
+            'rating' => 'required|integer|min:1|max:5',
+            'image' => 'nullable|string|max:255',
+            'order' => 'nullable|integer|min:0',
+            'is_visible' => 'nullable|boolean',
+        ];
+    }
+}